% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average-employment-rate.R
\name{average_employment_rate}
\alias{average_employment_rate}
\title{Calculate Average Employment Rate for a Major Category}
\usage{
average_employment_rate(data, major_category)
}
\arguments{
\item{data}{A data frame containing the dataset}

\item{major_category}{The major category for which you want to calculate the average employment rate.}
}
\value{
The average employment rate rounded to one decimal place.
}
\description{
This function calculates the average employment rate for a specific major category.
The employment rate is calculated as 100\% minus the unemployment rate.
}
\examples{
average_employment_rate(recent_grads, "Engineering")

}
