% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-female-grad-percent.R
\name{calc_female_grad_percent}
\alias{calc_female_grad_percent}
\title{Calculate Proportion of Female Graduates for a Major Category}
\usage{
calc_female_grad_percent(data, major_category)
}
\arguments{
\item{data}{A data frame containing the dataset}

\item{major_category}{The major category for which you want to calculate the proportion of female graduates.}
}
\value{
The percentage of female graduates rounded to one decimal place.
}
\description{
This function calculates the percentage of female graduates in a specific major category.
}
\examples{
calc_female_grad_percent(recent_grads, "Engineering")
calc_female_grad_percent(recent_grads, "Business")

}
