% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-gender-representation-plot.R
\name{generate_gender_representation_plot}
\alias{generate_gender_representation_plot}
\title{Generate a Gender Representation Plot for College Majors}
\usage{
generate_gender_representation_plot(major_category, selected_major, data)
}
\arguments{
\item{major_category}{The major category you want to analyze.}

\item{selected_major}{A character vector specifying one or more specific majors within the major category. Use "All" to include all majors in the category.}

\item{data}{A data frame containing the dataset.}
}
\value{
An interactive plot showing the gender distribution for the selected major(s) within the specified major category.
}
\description{
This function generates an interactive plot displaying the gender distribution for college majors within a specific major category.
}
\examples{
generate_gender_representation_plot("Arts", "All", recent_grads)
generate_gender_representation_plot("Engineering", c("Architecture", "Civil Engineering"), recent_grads)

}
