% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradmajors-package.R
\docType{package}
\name{gradmajors-package}
\alias{gradmajors}
\alias{gradmajors-package}
\title{gradmajors: U.S. College Majors Data Visualization and Analysis}
\description{
The 'gradmajors' package provides a collection of tools and functions to explore and analyze U.S. college majors data from 2010 to 2012. Uncover insights into employment rates, gender representation, and more in the American higher education landscape. Calculate average employment rates, assess the percentage of female graduates, create interactive gender distribution plots, and launch a Shiny application for immersive data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://651d0a6021358814d1ed92a9--rpkg-agoh0008-35d745.netlify.app/}
}

}
\author{
\strong{Maintainer}: Alexandra Goh \email{agoh0008@student.monash.edu}

}
\keyword{internal}
