% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recent-grads.R
\name{recent_grads}
\alias{recent_grads}
\alias{load_and_prepare_recent_grads}
\title{Load and Prepare the recent_grads Dataset}
\usage{
load_and_prepare_recent_grads(url)
}
\arguments{
\item{url}{The URL of the dataset to download.}
}
\value{
A cleaned and prepared data frame containing the \code{recent_grads} dataset.
}
\description{
This function loads and prepares the \code{recent_grads} dataset for analysis.
It downloads the dataset from the specified URL, reads it into a data frame,
and performs any necessary data cleaning or preprocessing.
}
\examples{
\dontrun{
# Load and prepare the dataset
recent_grads <- load_and_prepare_recent_grads(
  'https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2018/2018-10-16/recent-grads.csv'
)
}

}
