% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_app}
\alias{run_app}
\title{Run U.S. College Majors Shiny Application}
\usage{
run_app()
}
\value{
The Shiny application is launched in the default web browser.
}
\description{
This function launches a Shiny application that provides interactive visualizations
and analysis tools for U.S. college majors data (2010-2012).
}
\details{
The function locates the Shiny application directory within the "gradmajors" package and launches the application.
}
